+++ crates/starpls_hir/src/display.rs
@@ -306,6 +306,7 @@ impl DisplayWithDb for TyKind {
             TyKind::ModuleExtensionProxy(_) => "module_extension_proxy",
             TyKind::Tag(_) => "tag",
             TyKind::Target => "Target",
+            TyKind::Macro(_) => "macro",
         };
 
         f.write_str(text)
